pyproject.toml
[project]
name = "fact-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "exa-py>=2.7.0",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
